Refactor openssl code one more time before rustls integration
