+++ pingora-core/Cargo.toml
@@ -26,6 +26,7 @@ pingora-pool = { version = "0.3.0", path = "../pingora-pool" }
 pingora-error = { version = "0.3.0", path = "../pingora-error" }
 pingora-timeout = { version = "0.3.0", path = "../pingora-timeout" }
 pingora-http = { version = "0.3.0", path = "../pingora-http" }
+pingora-rustls = { version = "0.3.0", path = "../pingora-rustls", optional = true }
 tokio = { workspace = true, features = ["net", "rt-multi-thread", "signal"] }
 futures = "0.3"
 async-trait = { workspace = true }
@@ -66,6 +67,7 @@ openssl-probe = "0.1"
 tokio-test = "0.4"
 zstd = "0"
 httpdate = "1"
+x509-parser = { version = "0.16.0", optional = true }
 
 [target.'cfg(unix)'.dependencies]
 daemonize = "0.5.0"
@@ -87,9 +89,11 @@ hyperlocal = "0.8"
 jemallocator = "0.5"
 
 [features]
-default = ["openssl"]
-openssl = ["pingora-openssl", "some_tls"]
-boringssl = ["pingora-boringssl", "some_tls"]
+default = []
+openssl = ["pingora-openssl", "openssl_derived",]
+boringssl = ["pingora-boringssl", "openssl_derived",]
+rustls = ["pingora-rustls", "any_tls", "dep:x509-parser"]
+patched_http1 = ["pingora-http/patched_http1"]
+openssl_derived = ["any_tls"]
+any_tls = []
 sentry = ["dep:sentry"]
-patched_http1 = []
-some_tls = []
