+++ pingora-core/src/connectors/http/mod.rs
@@ -95,7 +95,7 @@ impl Connector {
 }
 
 #[cfg(test)]
-#[cfg(feature = "some_tls")]
+#[cfg(feature = "any_tls")]
 mod tests {
     use super::*;
     use crate::protocols::http::v1::client::HttpSession as Http1Session;
