+++ pingora-core/src/lib.rs
@@ -55,11 +55,14 @@ pub use pingora_error::{ErrorType::*, *};
 #[cfg(feature = "boringssl")]
 pub use pingora_boringssl as tls;
 
-#[cfg(all(not(feature = "boringssl"), feature = "openssl"))]
+#[cfg(feature = "openssl")]
 pub use pingora_openssl as tls;
 
-#[cfg(not(feature = "some_tls"))]
-pub use protocols::tls::dummy_tls as tls;
+#[cfg(feature = "rustls")]
+pub use pingora_rustls as tls;
+
+#[cfg(not(feature = "any_tls"))]
+pub use protocols::tls::noop_tls as tls;
 
 pub mod prelude {
     pub use crate::server::configuration::Opt;
