+++ pingora-core/src/listeners/mod.rs
@@ -15,21 +15,40 @@
 //! The listening endpoints (TCP and TLS) and their configurations.
 
 mod l4;
-mod tls;
 
-use crate::protocols::Stream;
+#[cfg(feature = "any_tls")]
+pub mod tls;
+
+#[cfg(not(feature = "any_tls"))]
+pub use crate::tls::listeners as tls;
+
+use crate::protocols::{tls::TlsRef, Stream};
+
 #[cfg(unix)]
 use crate::server::ListenFds;
 
+use async_trait::async_trait;
 use pingora_error::Result;
 use std::{fs::Permissions, sync::Arc};
 
 use l4::{ListenerEndpoint, Stream as L4Stream};
-use tls::Acceptor;
+use tls::{Acceptor, TlsSettings};
 
-pub use crate::protocols::tls::server::TlsAccept;
+pub use crate::protocols::tls::ALPN;
 pub use l4::{ServerAddress, TcpSocketOptions};
-pub use tls::{TlsSettings, ALPN};
+
+/// The APIs to customize things like certificate during TLS server side handshake
+#[async_trait]
+pub trait TlsAccept {
+    // TODO: return error?
+    /// This function is called in the middle of a TLS handshake. Structs who implement this function
+    /// should provide tls certificate and key to the [TlsRef] via [ext::ssl_use_certificate] and [ext::ssl_use_private_key].
+    async fn certificate_callback(&self, _ssl: &mut TlsRef) -> () {
+        // does nothing by default
+    }
+}
+
+pub type TlsAcceptCallbacks = Box<dyn TlsAccept + Send + Sync>;
 
 struct TransportStackBuilder {
     l4: ServerAddress,
@@ -200,6 +219,7 @@ impl Listeners {
 #[cfg(test)]
 mod test {
     use super::*;
+    #[cfg(feature = "any_tls")]
     use tokio::io::AsyncWriteExt;
     use tokio::net::TcpStream;
     use tokio::time::{sleep, Duration};
@@ -233,7 +253,7 @@ mod test {
     }
 
     #[tokio::test]
-    #[cfg(feature = "some_tls")]
+    #[cfg(feature = "any_tls")]
     async fn test_listen_tls() {
         use tokio::io::AsyncReadExt;
