+++ pingora-core/src/listeners/tls/boringssl_openssl/mod.rs
@@ -16,14 +16,16 @@ use log::debug;
 use pingora_error::{ErrorType, OrErr, Result};
 use std::ops::{Deref, DerefMut};
 
-use crate::protocols::tls::{
-    server::{handshake, handshake_with_callback, TlsAcceptCallbacks},
-    SslStream,
-};
+pub use crate::protocols::tls::ALPN;
 use crate::protocols::IO;
 use crate::tls::ssl::{SslAcceptor, SslAcceptorBuilder, SslFiletype, SslMethod};
-
-pub use crate::protocols::tls::ALPN;
+use crate::{
+    listeners::TlsAcceptCallbacks,
+    protocols::tls::{
+        server::{handshake, handshake_with_callback},
+        SslStream,
+    },
+};
 
 pub const TLS_CONF_ERR: ErrorType = ErrorType::Custom("TLSConfigError");
