+++ pingora-core/src/protocols/tls/boringssl_openssl/mod.rs
@@ -1,2 +1,29 @@
+// Copyright 2024 Cloudflare, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 pub mod client;
 pub mod server;
+mod stream;
+
+#[cfg(feature = "boringssl")]
+use pingora_boringssl as ssl_lib;
+
+#[cfg(feature = "openssl")]
+use pingora_openssl as ssl_lib;
+
+use ssl_lib::{ssl::SslRef, x509::X509};
+pub use stream::*;
+
+pub type TlsRef = SslRef;
+pub type CaType = Box<[X509]>;
