+++ pingora-core/src/protocols/tls/mod.rs
@@ -15,177 +15,19 @@
 //! The TLS layer implementations
 
 pub mod digest;
+pub use digest::*;
 
-#[cfg(feature = "some_tls")]
+#[cfg(feature = "openssl_derived")]
 mod boringssl_openssl;
 
-#[cfg(feature = "some_tls")]
+#[cfg(feature = "openssl_derived")]
 pub use boringssl_openssl::*;
 
-#[cfg(not(feature = "some_tls"))]
-pub mod dummy_tls;
+#[cfg(not(feature = "any_tls"))]
+pub mod noop_tls;
 
-use crate::protocols::digest::TimingDigest;
-use crate::protocols::{Peek, Ssl, UniqueID, UniqueIDType};
-use crate::tls::{self, ssl, tokio_ssl::SslStream as InnerSsl};
-use log::warn;
-use pingora_error::{ErrorType::*, OrErr, Result};
-use std::pin::Pin;
-use std::sync::Arc;
-use std::task::{Context, Poll};
-use std::time::SystemTime;
-use tokio::io::{self, AsyncRead, AsyncWrite, ReadBuf};
-
-pub use digest::SslDigest;
-
-/// The TLS connection
-#[derive(Debug)]
-pub struct SslStream<T> {
-    ssl: InnerSsl<T>,
-    digest: Option<Arc<SslDigest>>,
-    timing: TimingDigest,
-}
-
-impl<T> SslStream<T>
-where
-    T: AsyncRead + AsyncWrite + std::marker::Unpin,
-{
-    /// Create a new TLS connection from the given `stream`
-    ///
-    /// The caller needs to perform [`Self::connect()`] or [`Self::accept()`] to perform TLS
-    /// handshake after.
-    pub fn new(ssl: ssl::Ssl, stream: T) -> Result<Self> {
-        let ssl = InnerSsl::new(ssl, stream)
-            .explain_err(TLSHandshakeFailure, |e| format!("ssl stream error: {e}"))?;
-
-        Ok(SslStream {
-            ssl,
-            digest: None,
-            timing: Default::default(),
-        })
-    }
-
-    /// Connect to the remote TLS server as a client
-    pub async fn connect(&mut self) -> Result<(), ssl::Error> {
-        Self::clear_error();
-        Pin::new(&mut self.ssl).connect().await?;
-        self.timing.established_ts = SystemTime::now();
-        self.digest = Some(Arc::new(SslDigest::from_ssl(self.ssl())));
-        Ok(())
-    }
-
-    /// Finish the TLS handshake from client as a server
-    pub async fn accept(&mut self) -> Result<(), ssl::Error> {
-        Self::clear_error();
-        Pin::new(&mut self.ssl).accept().await?;
-        self.timing.established_ts = SystemTime::now();
-        self.digest = Some(Arc::new(SslDigest::from_ssl(self.ssl())));
-        Ok(())
-    }
-
-    #[inline]
-    fn clear_error() {
-        let errs = tls::error::ErrorStack::get();
-        if !errs.errors().is_empty() {
-            warn!("Clearing dirty TLS error stack: {}", errs);
-        }
-    }
-}
-
-impl<T> SslStream<T> {
-    pub fn ssl_digest(&self) -> Option<Arc<SslDigest>> {
-        self.digest.clone()
-    }
-}
-
-use std::ops::{Deref, DerefMut};
-
-impl<T> Deref for SslStream<T> {
-    type Target = InnerSsl<T>;
-
-    fn deref(&self) -> &Self::Target {
-        &self.ssl
-    }
-}
-
-impl<T> DerefMut for SslStream<T> {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        &mut self.ssl
-    }
-}
-
-impl<T> AsyncRead for SslStream<T>
-where
-    T: AsyncRead + AsyncWrite + Unpin,
-{
-    fn poll_read(
-        mut self: Pin<&mut Self>,
-        cx: &mut Context<'_>,
-        buf: &mut ReadBuf<'_>,
-    ) -> Poll<io::Result<()>> {
-        Self::clear_error();
-        Pin::new(&mut self.ssl).poll_read(cx, buf)
-    }
-}
-
-impl<T> AsyncWrite for SslStream<T>
-where
-    T: AsyncRead + AsyncWrite + Unpin,
-{
-    fn poll_write(
-        mut self: Pin<&mut Self>,
-        cx: &mut Context,
-        buf: &[u8],
-    ) -> Poll<io::Result<usize>> {
-        Self::clear_error();
-        Pin::new(&mut self.ssl).poll_write(cx, buf)
-    }
-
-    fn poll_flush(mut self: Pin<&mut Self>, cx: &mut Context) -> Poll<io::Result<()>> {
-        Self::clear_error();
-        Pin::new(&mut self.ssl).poll_flush(cx)
-    }
-
-    fn poll_shutdown(mut self: Pin<&mut Self>, cx: &mut Context) -> Poll<io::Result<()>> {
-        Self::clear_error();
-        Pin::new(&mut self.ssl).poll_shutdown(cx)
-    }
-
-    fn poll_write_vectored(
-        mut self: Pin<&mut Self>,
-        cx: &mut Context<'_>,
-        bufs: &[std::io::IoSlice<'_>],
-    ) -> Poll<io::Result<usize>> {
-        Self::clear_error();
-        Pin::new(&mut self.ssl).poll_write_vectored(cx, bufs)
-    }
-
-    fn is_write_vectored(&self) -> bool {
-        true
-    }
-}
-
-impl<T> UniqueID for SslStream<T>
-where
-    T: UniqueID,
-{
-    fn id(&self) -> UniqueIDType {
-        self.ssl.get_ref().id()
-    }
-}
-
-impl<T> Ssl for SslStream<T> {
-    fn get_ssl(&self) -> Option<&ssl::SslRef> {
-        Some(self.ssl())
-    }
-
-    fn get_ssl_digest(&self) -> Option<Arc<SslDigest>> {
-        self.ssl_digest()
-    }
-}
-
-// TODO: implement Peek if needed
-impl<T> Peek for SslStream<T> {}
+#[cfg(not(feature = "any_tls"))]
+pub use noop_tls::*;
 
 /// The protocol for Application-Layer Protocol Negotiation
 #[derive(Hash, Clone, Debug)]
@@ -236,6 +78,7 @@ impl ALPN {
         }
     }
 
+    #[cfg(feature = "openssl_derived")]
     pub(crate) fn to_wire_preference(&self) -> &[u8] {
         // https://www.openssl.org/docs/manmaster/man3/SSL_CTX_set_alpn_select_cb.html
         // "vector of nonempty, 8-bit length-prefixed, byte strings"
@@ -246,11 +89,21 @@ impl ALPN {
         }
     }
 
+    #[cfg(feature = "any_tls")]
     pub(crate) fn from_wire_selected(raw: &[u8]) -> Option<Self> {
         match raw {
             b"http/1.1" => Some(Self::H1),
             b"h2" => Some(Self::H2),
             _ => None,
+        }
+    }
+
+    #[cfg(feature = "rustls")]
+    pub(crate) fn to_wire_protocols(&self) -> Vec<Vec<u8>> {
+        match self {
+            ALPN::H1 => vec![b"http/1.1".to_vec()],
+            ALPN::H2 => vec![b"h2".to_vec()],
+            ALPN::H2H1 => vec![b"h2".to_vec(), b"http/1.1".to_vec()],
         }
     }
 }
