+++ pingora-error/src/lib.rs
@@ -105,6 +105,7 @@ pub enum ErrorType {
     ConnectTimedout,
     ConnectRefused,
     ConnectNoRoute,
+    TLSWantX509Lookup,
     TLSHandshakeFailure,
     TLSHandshakeTimedout,
     InvalidCert,
@@ -164,6 +165,7 @@ impl ErrorType {
             ErrorType::ConnectRefused => "ConnectRefused",
             ErrorType::ConnectNoRoute => "ConnectNoRoute",
             ErrorType::ConnectProxyFailure => "ConnectProxyFailure",
+            ErrorType::TLSWantX509Lookup => "TLSWantX509Lookup",
             ErrorType::TLSHandshakeFailure => "TLSHandshakeFailure",
             ErrorType::TLSHandshakeTimedout => "TLSHandshakeTimedout",
             ErrorType::InvalidCert => "InvalidCert",
