+++ pingora-load-balancing/Cargo.toml
@@ -35,6 +35,9 @@ derivative = "2.2.0"
 [dev-dependencies]
 
 [features]
-default = ["openssl"]
-openssl = ["pingora-core/openssl"]
-boringssl = ["pingora-core/boringssl"]
+default = []
+openssl = ["pingora-core/openssl", "openssl_derived"]
+boringssl = ["pingora-core/boringssl", "openssl_derived"]
+rustls = ["pingora-core/rustls", "any_tls"]
+openssl_derived = ["any_tls"]
+any_tls = []
