+++ pingora-proxy/Cargo.toml
@@ -44,7 +44,7 @@ env_logger = "0.9"
 hyper = "0.14"
 tokio-tungstenite = "0.20.1"
 pingora-limits = { version = "0.3.0", path = "../pingora-limits" }
-pingora-load-balancing = { version = "0.3.0", path = "../pingora-load-balancing" }
+pingora-load-balancing = { version = "0.3.0", path = "../pingora-load-balancing", default-features=false }
 prometheus = "0"
 futures-util = "0.3"
 serde = { version = "1.0", features = ["derive"] }
@@ -55,9 +55,12 @@ serde_yaml = "0.8"
 hyperlocal = "0.8"
 
 [features]
-default = ["openssl"]
-openssl = ["pingora-core/openssl", "pingora-cache/openssl"]
-boringssl = ["pingora-core/boringssl", "pingora-cache/boringssl"]
+default = []
+openssl = ["pingora-core/openssl", "pingora-cache/openssl", "openssl_derived"]
+boringssl = ["pingora-core/boringssl", "pingora-cache/boringssl", "openssl_derived"]
+rustls = ["pingora-core/rustls", "pingora-cache/rustls", "any_tls"]
+openssl_derived = ["any_tls"]
+any_tls = []
 sentry = ["pingora-core/sentry"]
 
 # or locally cargo doc --config "build.rustdocflags='--cfg doc_async_trait'"
