+++ pingora-proxy/tests/test_basic.rs
@@ -67,6 +67,7 @@ async fn test_simple_proxy() {
 }
 
 #[tokio::test]
+#[cfg(feature = "any_tls")]
 async fn test_h2_to_h1() {
     init();
     let client = reqwest::Client::builder()
@@ -74,7 +75,12 @@ async fn test_h2_to_h1() {
         .build()
         .unwrap();
 
-    let res = client.get("https://127.0.0.1:6150").send().await.unwrap();
+    let res = client
+        .get("https://127.0.0.1:6150")
+        .header("sni", "openrusty.org")
+        .send()
+        .await
+        .unwrap();
     assert_eq!(res.status(), reqwest::StatusCode::OK);
     assert_eq!(res.version(), reqwest::Version::HTTP_2);
 
@@ -104,6 +110,7 @@ async fn test_h2_to_h1() {
 }
 
 #[tokio::test]
+#[cfg(feature = "any_tls")]
 async fn test_h2_to_h2() {
     init();
     let client = reqwest::Client::builder()
@@ -113,6 +120,7 @@ async fn test_h2_to_h2() {
 
     let res = client
         .get("https://127.0.0.1:6150")
+        .header("sni", "openrusty.org")
         .header("x-h2", "true")
         .send()
         .await
@@ -189,6 +197,7 @@ async fn test_h1_on_h2c_port() {
 }
 
 #[tokio::test]
+#[cfg(feature = "openssl_derived")]
 async fn test_h2_to_h2_host_override() {
     init();
     let client = reqwest::Client::builder()
@@ -212,6 +221,7 @@ async fn test_h2_to_h2_host_override() {
 }
 
 #[tokio::test]
+#[cfg(feature = "any_tls")]
 async fn test_h2_to_h2_upload() {
     init();
     let client = reqwest::Client::builder()
@@ -223,6 +233,7 @@ async fn test_h2_to_h2_upload() {
 
     let res = client
         .get("https://127.0.0.1:6150/echo")
+        .header("sni", "openrusty.org")
         .header("x-h2", "true")
         .body(payload)
         .send()
@@ -235,6 +246,7 @@ async fn test_h2_to_h2_upload() {
 }
 
 #[tokio::test]
+#[cfg(feature = "any_tls")]
 async fn test_h2_to_h1_upload() {
     init();
     let client = reqwest::Client::builder()
@@ -246,6 +258,7 @@ async fn test_h2_to_h1_upload() {
 
     let res = client
         .get("https://127.0.0.1:6150/echo")
+        .header("sni", "openrusty.org")
         .body(payload)
         .send()
         .await
@@ -313,7 +326,10 @@ async fn test_simple_proxy_uds_peer() {
     assert!(is_specified_port(sockaddr.port()));
 
     assert_eq!(headers["x-upstream-client-addr"], "unset"); // unnamed UDS
-    assert_eq!(headers["x-upstream-server-addr"], "/tmp/nginx-test.sock");
+    assert_eq!(
+        headers["x-upstream-server-addr"],
+        "/tmp/pingora_nginx_test.sock"
+    );
 
     let body = res.text().await.unwrap();
     assert_eq!(body, "Hello World!\n");
@@ -444,6 +460,8 @@ async fn test_dropped_conn() {
     test_dropped_conn_post_body_over().await;
 }
 
+// currently not supported with Rustls implementation
+#[cfg(feature = "openssl_derived")]
 #[tokio::test]
 async fn test_tls_no_verify() {
     init();
@@ -457,6 +475,7 @@ async fn test_tls_no_verify() {
     assert_eq!(res.status(), StatusCode::OK);
 }
 
+#[cfg(feature = "any_tls")]
 #[tokio::test]
 async fn test_tls_verify_sni_not_host() {
     init();
@@ -473,6 +492,8 @@ async fn test_tls_verify_sni_not_host() {
     assert_eq!(res.status(), StatusCode::OK);
 }
 
+// currently not supported with Rustls implementation
+#[cfg(feature = "openssl_derived")]
 #[tokio::test]
 async fn test_tls_none_verify_host() {
     init();
@@ -489,6 +510,7 @@ async fn test_tls_none_verify_host() {
     assert_eq!(res.status(), StatusCode::OK);
 }
 
+#[cfg(feature = "any_tls")]
 #[tokio::test]
 async fn test_tls_verify_sni_host() {
     init();
@@ -506,6 +528,7 @@ async fn test_tls_verify_sni_host() {
     assert_eq!(res.status(), StatusCode::OK);
 }
 
+#[cfg(feature = "any_tls")]
 #[tokio::test]
 async fn test_tls_underscore_sub_sni_verify_host() {
     init();
@@ -523,6 +546,7 @@ async fn test_tls_underscore_sub_sni_verify_host() {
     assert_eq!(res.status(), StatusCode::OK);
 }
 
+#[cfg(feature = "any_tls")]
 #[tokio::test]
 async fn test_tls_underscore_non_sub_sni_verify_host() {
     init();
@@ -542,6 +566,7 @@ async fn test_tls_underscore_non_sub_sni_verify_host() {
     assert_eq!(headers[header::CONNECTION], "close");
 }
 
+#[cfg(feature = "any_tls")]
 #[tokio::test]
 async fn test_tls_alt_verify_host() {
     init();
@@ -560,6 +585,7 @@ async fn test_tls_alt_verify_host() {
     assert_eq!(res.status(), StatusCode::OK);
 }
 
+#[cfg(feature = "any_tls")]
 #[tokio::test]
 async fn test_tls_underscore_sub_alt_verify_host() {
     init();
@@ -578,6 +604,7 @@ async fn test_tls_underscore_sub_alt_verify_host() {
     assert_eq!(res.status(), StatusCode::OK);
 }
 
+#[cfg(feature = "any_tls")]
 #[tokio::test]
 async fn test_tls_underscore_non_sub_alt_verify_host() {
     init();
@@ -691,6 +718,7 @@ async fn test_connect_close() {
 }
 
 #[tokio::test]
+#[cfg(feature = "any_tls")]
 async fn test_mtls_no_client_cert() {
     init();
     let client = reqwest::Client::new();
@@ -709,6 +737,7 @@ async fn test_mtls_no_client_cert() {
     assert_eq!(res.status(), StatusCode::BAD_REQUEST);
 }
 
+#[cfg(feature = "any_tls")]
 #[tokio::test]
 async fn test_mtls_no_intermediate_cert() {
     init();
@@ -730,6 +759,7 @@ async fn test_mtls_no_intermediate_cert() {
 }
 
 #[tokio::test]
+#[cfg(feature = "any_tls")]
 async fn test_mtls() {
     init();
     let client = reqwest::Client::new();
@@ -748,6 +778,7 @@ async fn test_mtls() {
     assert_eq!(res.status(), StatusCode::OK);
 }
 
+#[cfg(feature = "any_tls")]
 async fn assert_reuse(req: reqwest::RequestBuilder) {
     req.try_clone().unwrap().send().await.unwrap();
     let res = req.send().await.unwrap();
@@ -755,6 +786,7 @@ async fn assert_reuse(req: reqwest::RequestBuilder) {
     assert!(headers.get("x-conn-reuse").is_some());
 }
 
+#[cfg(feature = "any_tls")]
 #[tokio::test]
 async fn test_mtls_diff_cert_no_reuse() {
     init();
@@ -789,6 +821,7 @@ async fn test_mtls_diff_cert_no_reuse() {
     assert!(headers.get("x-conn-reuse").is_none());
 }
 
+#[cfg(feature = "any_tls")]
 #[tokio::test]
 async fn test_tls_diff_verify_no_reuse() {
     init();
@@ -815,6 +848,7 @@ async fn test_tls_diff_verify_no_reuse() {
     assert!(headers.get("x-conn-reuse").is_none());
 }
 
+#[cfg(feature = "any_tls")]
 #[tokio::test]
 async fn test_tls_diff_verify_host_no_reuse() {
     init();
@@ -843,6 +877,7 @@ async fn test_tls_diff_verify_host_no_reuse() {
     assert!(headers.get("x-conn-reuse").is_none());
 }
 
+#[cfg(feature = "any_tls")]
 #[tokio::test]
 async fn test_tls_diff_alt_cnt_no_reuse() {
     init();
