+++ pingora-proxy/tests/utils/conf/origin/conf/nginx.conf
@@ -6,7 +6,7 @@ error_log  /dev/stdout;
 #error_log  logs/error.log  notice;
 #error_log  logs/error.log  info;
 
-pid        /tmp/mock_origin.pid;
+pid        /tmp/pingora_mock_origin.pid;
 master_process off;
 daemon off;
 
@@ -85,7 +85,7 @@ http {
         listen       8001;
         listen       [::]:8000;
         #listen       8443 ssl;
-        listen       unix:/tmp/nginx-test.sock;
+        listen       unix:/tmp/pingora_nginx_test.sock;
         listen       8443 ssl http2;
         server_name  localhost;
 
@@ -97,6 +97,9 @@ http {
         # for benchmark
         http2_max_requests 999999;
 
+        # increase max body size for /upload/ test
+        client_max_body_size 128m;
+        
         #charset koi8-r;
 
         #access_log  logs/host.access.log  main;
