+++ pingora-proxy/tests/utils/mod.rs
@@ -14,7 +14,9 @@
 
 #![allow(unused)]
 
+#[cfg(feature = "any_tls")]
 pub mod cert;
+
 pub mod mock_origin;
 pub mod server_utils;
 pub mod websocket;
