+++ pingora-proxy/tests/utils/server_utils.rs
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#[cfg(feature = "any_tls")]
 use super::cert;
 use async_trait::async_trait;
 use clap::Parser;
@@ -32,7 +33,7 @@ use pingora_core::protocols::{l4::socket::SocketAddr, Digest};
 use pingora_core::server::configuration::Opt;
 use pingora_core::services::Service;
 use pingora_core::upstreams::peer::HttpPeer;
-use pingora_core::utils::CertKey;
+use pingora_core::utils::tls::CertKey;
 use pingora_error::{Error, ErrorSource, Result};
 use pingora_http::{RequestHeader, ResponseHeader};
 use pingora_proxy::{ProxyHttp, Session};
@@ -106,6 +107,7 @@ fn response_filter_common(
 }
 
 #[async_trait]
+#[cfg(feature = "any_tls")]
 impl ProxyHttp for ExampleProxyHttps {
     type CTX = CTX;
     fn new_ctx(&self) -> Self::CTX {
@@ -283,7 +285,7 @@ impl ProxyHttp for ExampleProxyHttp {
         #[cfg(unix)]
         if req.headers.contains_key("x-uds-peer") {
             return Ok(Box::new(HttpPeer::new_uds(
-                "/tmp/nginx-test.sock",
+                "/tmp/pingora_nginx_test.sock",
                 false,
                 "".to_string(),
             )?));
@@ -558,26 +560,35 @@ fn test_main() {
     http_logic.server_options = Some(http_server_options);
     proxy_service_h2c.add_tcp("0.0.0.0:6146");
 
-    let mut proxy_service_https =
-        pingora_proxy::http_proxy_service(&my_server.configuration, ExampleProxyHttps {});
-    proxy_service_https.add_tcp("0.0.0.0:6149");
-    let cert_path = format!("{}/tests/keys/server.crt", env!("CARGO_MANIFEST_DIR"));
-    let key_path = format!("{}/tests/keys/key.pem", env!("CARGO_MANIFEST_DIR"));
-    let mut tls_settings =
-        pingora_core::listeners::TlsSettings::intermediate(&cert_path, &key_path).unwrap();
-    tls_settings.enable_h2();
-    proxy_service_https.add_tls_with_settings("0.0.0.0:6150", None, tls_settings);
+    let mut proxy_service_https_opt: Option<Box<dyn Service>> = None;
+
+    #[cfg(feature = "any_tls")]
+    {
+        let mut proxy_service_https =
+            pingora_proxy::http_proxy_service(&my_server.configuration, ExampleProxyHttps {});
+        proxy_service_https.add_tcp("0.0.0.0:6149");
+        let cert_path = format!("{}/tests/keys/server.crt", env!("CARGO_MANIFEST_DIR"));
+        let key_path = format!("{}/tests/keys/key.pem", env!("CARGO_MANIFEST_DIR"));
+        let mut tls_settings =
+            pingora_core::listeners::tls::TlsSettings::intermediate(&cert_path, &key_path).unwrap();
+        tls_settings.enable_h2();
+        proxy_service_https.add_tls_with_settings("0.0.0.0:6150", None, tls_settings);
+        proxy_service_https_opt = Some(Box::new(proxy_service_https))
+    }
 
     let mut proxy_service_cache =
         pingora_proxy::http_proxy_service(&my_server.configuration, ExampleProxyCache {});
     proxy_service_cache.add_tcp("0.0.0.0:6148");
 
-    let services: Vec<Box<dyn Service>> = vec![
+    let mut services: Vec<Box<dyn Service>> = vec![
         Box::new(proxy_service_h2c),
         Box::new(proxy_service_http),
-        Box::new(proxy_service_https),
         Box::new(proxy_service_cache),
     ];
+
+    if let Some(proxy_service_https) = proxy_service_https_opt {
+        services.push(proxy_service_https)
+    }
 
     set_compression_dict_path("tests/headers.dict");
     my_server.add_services(services);
