+++ pingora-rustls/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "pingora-rustls"
+version = "0.3.0"
+authors = ["Harald Gutmann <harald@gutmann.one>"]
+license = "Apache-2.0"
+edition = "2021"
+repository = "https://github.com/cloudflare/pingora"
+categories = ["asynchronous", "network-programming"]
+keywords = ["async", "tls", "ssl", "pingora"]
+description = """
+RusTLS async APIs for Pingora.
+"""
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[lib]
+name = "pingora_rustls"
+path = "src/lib.rs"
