+++ pingora/examples/server.rs
@@ -15,6 +15,7 @@
 #[global_allocator]
 static GLOBAL: jemallocator::Jemalloc = jemallocator::Jemalloc;
 
+use pingora::listeners::tls::TlsSettings;
 use pingora::protocols::TcpKeepalive;
 use pingora::server::configuration::Opt;
 use pingora::server::{Server, ShutdownWatch};
@@ -49,31 +50,35 @@ impl BackgroundService for ExampleBackgroundService {
         }
     }
 }
+#[cfg(feature = "openssl_derived")]
+mod boringssl_openssl {
+    use super::*;
+    use pingora::tls::pkey::{PKey, Private};
+    use pingora::tls::x509::X509;
 
-use pingora::tls::pkey::{PKey, Private};
-use pingora::tls::x509::X509;
-struct DynamicCert {
-    cert: X509,
-    key: PKey<Private>,
-}
-
-impl DynamicCert {
-    fn new(cert: &str, key: &str) -> Box<Self> {
-        let cert_bytes = std::fs::read(cert).unwrap();
-        let cert = X509::from_pem(&cert_bytes).unwrap();
-
-        let key_bytes = std::fs::read(key).unwrap();
-        let key = PKey::private_key_from_pem(&key_bytes).unwrap();
-        Box::new(DynamicCert { cert, key })
+    pub(super) struct DynamicCert {
+        cert: X509,
+        key: PKey<Private>,
     }
-}
 
-#[async_trait]
-impl pingora::listeners::TlsAccept for DynamicCert {
-    async fn certificate_callback(&self, ssl: &mut pingora::tls::ssl::SslRef) {
-        use pingora::tls::ext;
-        ext::ssl_use_certificate(ssl, &self.cert).unwrap();
-        ext::ssl_use_private_key(ssl, &self.key).unwrap();
+    impl DynamicCert {
+        pub(super) fn new(cert: &str, key: &str) -> Box<Self> {
+            let cert_bytes = std::fs::read(cert).unwrap();
+            let cert = X509::from_pem(&cert_bytes).unwrap();
+
+            let key_bytes = std::fs::read(key).unwrap();
+            let key = PKey::private_key_from_pem(&key_bytes).unwrap();
+            Box::new(DynamicCert { cert, key })
+        }
+    }
+
+    #[async_trait]
+    impl pingora::listeners::TlsAccept for DynamicCert {
+        async fn certificate_callback(&self, ssl: &mut pingora::tls::ssl::SslRef) {
+            use pingora::tls::ext;
+            ext::ssl_use_certificate(ssl, &self.cert).unwrap();
+            ext::ssl_use_private_key(ssl, &self.key).unwrap();
+        }
     }
 }
 
@@ -132,12 +137,32 @@ pub fn main() {
     echo_service_http.add_tcp_with_settings("0.0.0.0:6145", options);
     echo_service_http.add_uds("/tmp/echo.sock", None);
 
-    let dynamic_cert = DynamicCert::new(&cert_path, &key_path);
-    let mut tls_settings = pingora::listeners::TlsSettings::with_callbacks(dynamic_cert).unwrap();
-    // by default intermediate supports both TLS 1.2 and 1.3. We force to tls 1.2 just for the demo
-    tls_settings
-        .set_max_proto_version(Some(pingora::tls::ssl::SslVersion::TLS1_2))
-        .unwrap();
+    let mut tls_settings;
+
+    // NOTE: dynamic certificate callback is only supported with BoringSSL/OpenSSL
+    #[cfg(feature = "openssl_derived")]
+    {
+        use std::ops::DerefMut;
+
+        let dynamic_cert = boringssl_openssl::DynamicCert::new(&cert_path, &key_path);
+        tls_settings = TlsSettings::with_callbacks(dynamic_cert).unwrap();
+        // by default intermediate supports both TLS 1.2 and 1.3. We force to tls 1.2 just for the demo
+
+        tls_settings
+            .deref_mut()
+            .deref_mut()
+            .set_max_proto_version(Some(pingora::tls::ssl::SslVersion::TLS1_2))
+            .unwrap();
+    }
+    #[cfg(feature = "rustls")]
+    {
+        tls_settings = TlsSettings::intermediate(&cert_path, &key_path).unwrap();
+    }
+    #[cfg(not(feature = "any_tls"))]
+    {
+        tls_settings = TlsSettings;
+    }
+
     tls_settings.enable_h2();
     echo_service_http.add_tls_with_settings("0.0.0.0:6148", None, tls_settings);
